(doc) add npm install as last step
For me, copying the files from the newly generated strapi environment was not sufficient. When I started the strapi after following the guide, 
I got the following error: 
```
/home/xxx/Dev/Projects/xxx/server/node_modules/mongodb/lib/mongo_client.js:804
          throw err;
          ^

TypeError: Cannot read property 'globalId' of undefined

```
I guess this was the case because strapi-mongoose was not upgraded.  I don't know if I'm blind or did not read the migration guide carefully enough, but for me a `npm install` at the end did the trick and upgraded it.
Maybe this is helpful for you