GraphQL many-to-many relations symmetry
Querying content types that share many-to-many relations do not return expected results.

Let's say we have a product content type and a category centent type. those two models share many-to-many relation, and when we are querying categories we have the expected result, Ex :

Query : 
{
  categories {
    Name
    products {
      Name
    }
  }
}

Result : 

{
  "data": {
    "categories": [
      {
        "Name": "Category 1",
        "products": [
          {
            "Name": "Product 1"
          }
        ]
      },
      {
        "Name": "Category 2",
        "products": [
          {
            "Name": "Product 1"
          }
        ]
      }
    ]
  }
}

But when  we're querying products, the categories array is empty!!

Query:

{
  products {
    Name
    categories {
      Name
    }
  }
}

Result:

{
  "data": {
    "products": [
      {
        "Name": "Product 1",
        "categories": [ ]
      }
    ]
  }
}

categories should not be empty, it should be something like  that : 
...
"categories": [
          {
            "Name": "Category 1"
          },
          {
            "Name": "Category 2"
          }
        ]
....

I hope that was clear.