Bugfix - Plugin Quick Start does not work as expected
If a user follows the directions in the plugin documentation quick start, several errors will show up, as documented in:

https://github.com/strapi/strapi/issues/1116
https://github.com/strapi/strapi/issues/1972

This updates the documentation so that a user new to the project will be able to have success. It follows the comments from https://github.com/strapi/strapi/issues/1972#issuecomment-422321650 & https://github.com/strapi/strapi/issues/1116#issuecomment-386710389 to re-write the docs to match expected behavior.