Fetch Github profile email if not found
When a Github user does not have a public email on his profile the
GitHub OAuth API will have `email` set to `null`. When the registered
OAuth application in Github has permission for the user:email scope it
is possible to get the user email via the `user/email` API endpoint.

This is similar to the `passport-github2` package.
See https://github.com/cfsghost/passport-github/blob/master/lib/strategy.js