Add rebuild after wysiwyg plugin has been created
Me and my colleague just tried this step by step guide to install the CKEditor WYSIWYG with the current strapi version 3.0.0-beta.20.1 and it didn't work.
We had to spend some time till we found out that we have to rebuild strapi so the plugin is loaded correctly.
Found that on the "Update Strapi version" page.

Signed-off-by: Kevin Pfeifer <kevin.pfeifer@sunlime.at>