Move click action to wrapper and update styles
Move click action to wrapper (parent) to allow clicks on
whole area of checkbox. Also update the highlighting and
visibility on hover of configuration wheel for better
usability. This way the user learns that clicking on the
item or wheel opens the configuration and clicking on
the text or checkbox toggles the checkbox.

Signed-off-by: Christos Malliaridis <c.malliaridis@gmail.com>