Updated Models Page Grammar (#7912)
* Updated Models Page Grammar

Quickly went through the models' page in the documentation and updated some grammar as well as added some emphasis and brief clarification in a couple of places.

* Edited slight typo.

* Updated Models Page Grammar

Quickly went through the models' page in the documentation and updated some grammar as well as added some emphasis and brief clarification in a couple of places.

Signed-off-by: Christopher Wray <53663762+cwray-tech@users.noreply.github.com>

* Edited slight typo.

Signed-off-by: Christopher Wray <53663762+cwray-tech@users.noreply.github.com>