fix: <svg> attribute width="<N>rem" errors in Safari (#21550)
* fix: <svg> attribute width="2rem" error in Safari
I changed the with and height value for all svgs used in the LeftMenu. It was changed from 2rem to 20px. Rem was leading to the error message Error: Invalid value for <svg> attribute width="2rem" in the safari browser.

* fix: <svg> attribute width="1.2rem" error in Safari
I changed the with and height value for all svgs used in the Onboarding. It was changed from 1.2rem to 12px. Rem was leading to the error message Error: Invalid value for attribute width="1.2rem" in the safari browser.

* Update packages/core/admin/admin/src/components/LeftMenu.tsx

Co-authored-by: Simone <startae14@gmail.com>

* 12px => 12

---------

Co-authored-by: Simone <startae14@gmail.com>